Sort mixin output in order declared (#2472)
* Sort mixin output in order declared

When we are querying mixins, such as when we are asking each mixin for Dockerfile lines to include in the bundle image, we call them in go routines. Depending on how quickly they return, the order is not guaranteed and may change. I have updated our mixin query function to sort the results of querying the mixins in a consistent order, the order that they are defined in the bundle.

The Execute function now returns a bit more information about the results of each mixin command, and it's up to the caller to decide how to handle errors when not all mixins are required to have a successful exit code. Lint allows mixins to not implement that command, so in that case we check if the command isn't implemented and skip over it.

Fixes #2469

During build in particular, this ensures that we optimize layer caching as much as possible by always outputing the same contents of the Dockerfile when building a bundle. Previously, it could include the mixins in varying order which prevented reuse of layers from previous builds.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Remove unused variable

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix race condition in unit test

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Make vet happy

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix capturing the error when lint isn't supported

The "unsupported command" text is printed to stderr. I have updated our mixin/plugin runner to capture the command output and include it in the returned error when the command fails.
That way we can check if a mixin command failed because the command isn't implemented.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Incorporate review feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>