Update sanitizer logic to allow for non-value based parameter overrides
The sanitizer assumed that all param overrides on an installation were hard coded values because that's what the CLI allows. With recent changes to support dependencies and workflows, internal parameter sets can now resolve secrets from other sources such as env, file, etc.

Update the sanitizer to not assume they are all values that should be sanitized or edited.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>