Do not print unsupported mixin command errors
Not all of the mixin commands are required, for example schema and lint are optional. Porter does ignore failed commands when the mixin doesn't support a called command, but after I updated that bit of code to use spans and strutured logging, I accidentally caused porter to always print out if the command failed, even when the verbosity is greater than debug.

This fixes failed mixin commands to only print the error when verbosity is debug and I've added a CLI test to validate that we print when expected to avoid future regressions in this area.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>