Update logger on context when finished loading configuration
We create a TraceLogger and place it into the context pretty early in porter's execution, before we are fully done applying flags to the confiugration. Later after we create a new TraceLogger with the proper configuration (such as verbosity), we weren't updating it on the context, so even when you ran a porter command with --verbosity=debug, it would use whatever was configured in the config file.

I've updated how we load the config so that it returns an updated context with the logger replaced. Then when functions lookup the TraceLogger from the context, they can get the correct instance.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>