Add guard rails to resolving the bundle reference
Avoid null reference exceptions by resolving and caching the bundle reference on the options, and not letting code directly access the "private" field that stored the resolved reference. Now we will resolve JIT if needed and cache the result for later use.

Ideally we would never have put state on the options in the first place but here we are. At least this will avoid accidental null references.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>