Consolidate modifying the installation based on opts into applyActionOptionsToInstallation
Instead of setting the modified timestamp, tracking the bundle reference, etc individually in the install/upgrade/uninstall, etc actions, I've consolidated them into applyActionOptionsToInstallation, and ensure that we validate the installation after modifying it. Now we only need to validate in this function before saving the record.

The unit tests for IsInstallationInSync made a lot of simplifying assumptions that let it do less test setup that I have addressed so that it's safe to call the function under test without calling it from inside the install/upgade/uninstall/invoke actions which handle prepping the installation and options first.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>