Refactor param resolution (#2476)
* Update sanitizer logic to allow for non-value based parameter overrides

The sanitizer assumed that all param overrides on an installation were hard coded values because that's what the CLI allows. With recent changes to support dependencies and workflows, internal parameter sets can now resolve secrets from other sources such as env, file, etc.

Update the sanitizer to not assume they are all values that should be sanitized or edited.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Improve applying installation parameters without reparsing as flags

Do not re-parse parameters set on an installation resource as --param flags and instead apply any --param flag overrides on top of parameters defined on the installation. This changes the current behavior which would remove any parameter overrides set on an installation resource when the overrides were not re-specified with --param.

Example:

```
porter install mybuns --param logLevel=10

porter installation show mybuns -o yaml > mybuns.yaml

porter installation apply mybuns.yaml
```

A few other improvements are:
* We only resolve parameters once, regardless of code path.
* We persist the installation from a dry-run after comparing state now and only persist if modified.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add guard rails to resolving the bundle reference

Avoid null reference exceptions by resolving and caching the bundle reference on the options, and not letting code directly access the "private" field that stored the resolved reference. Now we will resolve JIT if needed and cache the result for later use.

Ideally we would never have put state on the options in the first place but here we are. At least this will avoid accidental null references.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Always sanitize the installation file

I've combined applying options to an installation with calling sanitize so that it's easier to remember and never get accidentally skipped.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Do not modify the param varible in the sanitizer for loop

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Consolidate modifying the installation based on opts into applyActionOptionsToInstallation

Instead of setting the modified timestamp, tracking the bundle reference, etc individually in the install/upgrade/uninstall, etc actions, I've consolidated them into applyActionOptionsToInstallation, and ensure that we validate the installation after modifying it. Now we only need to validate in this function before saving the record.

The unit tests for IsInstallationInSync made a lot of simplifying assumptions that let it do less test setup that I have addressed so that it's safe to call the function under test without calling it from inside the install/upgade/uninstall/invoke actions which handle prepping the installation and options first.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

---------

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>