Consolidate testdata so we don't have a bunch of similar bundles
mybuns = a test bundle that contains as many features / test cases we can squeeze into a bundle, without making a bunch of paramters required so that it's easy to run. Long term I want to remove dependencies from this bundle though and switch to using a new set of bundles just for dependencies.

myenv = a metabundle that is comprised of its dependencies. Has a deep dependency graph.
myinfra = a bundle that creates infrastructure and generates outputs that parent bundles would want to consume.
myapp = a bundle that deploys an application and requires a db connection string
mydb = a leaf bundle that creates a db and outputs a connection string

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>