Remove name field from CNAB dependency representation in bundle.json
The name field for a dependency is not in the spec because CNAB represents a dependency as a map entry, keyed by name. We use a name internally just to make things easier while working with dependency entries but they shouldn't be written out with a name field.

I have updated how we serialize the dependency structs so that even when name is populated, we don't write it to bundle.json.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>