Keep support for multiple versions of a resource schemaVersion
Make it possible for Porter to support a range of schemaVersion for any resource, e.g. we support Installation 1.0.2-1.0.5. Right now we don't take advantage of that but since I wrote the logic and refactored a bunch of our resources (for a change that I've since reverted) I figured it was worth keeping so that we don't need to reimplement later.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>