Always set schemaType before persisting a resource
When the user applies a resource that does not have the schemaType set, we should default the field to the appropriate type, e.g. Installation/CredentialSpec, so that it's consistently set in the database. We don't use that field right now in Porter other than for validation but this ensures that it's consistent later when we may need to rely upon it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>