Split out file format docs per schemaVersion
Previously we continuously updated a single page with the most recent schema for each resource, making it hard to see the schema for a particular version. I've given each versino it's own page that calls out the changes and links directly to the json schema for that specific version. The index page will continue to show the most recent version.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>