Consistently set and use SchemaType when exporting/importing (#2428)
* Only use filesystem instead of a full porter context in the yq editor

I wanted to use the yaml editor helper in our regular tests too, not just the smoke test framework. So I tweaked the yq editor to only take in a filesystem and not require an entire porter context to avoid a circular dependency between portercontext and yaml editor package.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Consistently set and validate schemaType on resources

All resource files should have both a schemaVersion and a schemaType. The schemaType field serves multiple purposes:
* Porter can validate it when present during the `porter * apply` commands to catch if the wrong file type was passed
* In the future VS Code can provide autocomplete based on the schemaType and schemaVersion

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Document that bundle manifests now have schemaType

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Default SchemaType when a resource is validated

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Alias cnab-go schema.Version to cnab.SchemaVersion

This was missed when I originally did a sweep to isolate use of cnab-go to the cnab package. We want to avoid directly using cnab-go throughout Porter and give ourselves a single point to make adjustments or swap out types/functionality.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Keep support for multiple versions of a resource schemaVersion

Make it possible for Porter to support a range of schemaVersion for any resource, e.g. we support Installation 1.0.2-1.0.5. Right now we don't take advantage of that but since I wrote the logic and refactored a bunch of our resources (for a change that I've since reverted) I figured it was worth keeping so that we don't need to reimplement later.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Always set schemaType before persisting a resource

When the user applies a resource that does not have the schemaType set, we should default the field to the appropriate type, e.g. Installation/CredentialSpec, so that it's consistently set in the database. We don't use that field right now in Porter other than for validation but this ensures that it's consistent later when we may need to rely upon it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Split out file format docs per schemaVersion

Previously we continuously updated a single page with the most recent schema for each resource, making it hard to see the schema for a particular version. I've given each versino it's own page that calls out the changes and links directly to the json schema for that specific version. The index page will continue to show the most recent version.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

---------

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>