Fix associated Porter version for 1.0.1 manifest schema
Originally we planned to release v1.0.1 of the Porter manifest schema in Porter 1.10.0 but have since released it in v1.0.8. This updates the docs for the manifest to reflect that.

I have also created a new page for 1.0.1 of the manifest to match how the other file format pages work (with an index page and a separate page just for the current version).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>