Improve passing custom values as build arguments
Currently porter passes all custom values defined in porter.yaml as build arguments when using docker to build the invocation image.

```yaml
custom:
  app:
    version: 1.2.3
```

The above porter.yaml results in porter passing --build-arg CUSTOM_APP_VERSION=1.2.3

We have some users who are storing longer values in the custom map and have managed to hit the max argument length on their system. To address this, I am no longer passing all custom values as build arguments, and only pass them when the Dockerfile declares a build argument with the same name. In the example above, we only pass the app version when the Dockerfile declares `ARG CUSTOM_APP_VERSION`. In this users's case that's enough to avoid the problem.

In addition to the change above, I have also added a max length check on build arguments so that we can detect that it may be too long and provide a better error message. When we let the large value get passed to docker the error is simply "exit code 1" and we don't get any context into what failed. This will ensure that long values are caught before calling docker and telling the user to save the value in a file, and then read it back out in the Dockerfile instead.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>