Inject claim and not run
We were accidentally injecting Porter's run represetation, which is _very_ similar to the CNAB claim but isn't exactly the same. I've updated the code to add a test to ensure that we don't break injecting the claim.

For context, Porter doesn't use the injected claim but it is an optional part of the CNAB Claims spec, so while unused, if we are going to say we support it we should do it to spec.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>