Only set relevant parameter overrides on a run
An installation keeps track of previously used parameter overrides over time. Some parameters may apply to only specific actions. When a run is executed and some of the parameters remembered by the installation are not relevant to that run (for example, a parameter only applies to install and we are running upgrade), those irrelevant parameters were previously being persisted on the run regardless. I have fixed the code for creating a new run for an installation to only copy over parameters that apply to the run's action.

When this isn't in place, we can end up in a spot where sensitive parameters that don't apply to the current action are prepared to be stored in a secret, but never stored (because they don't apply) but later on we try to resolve all parameters associated with the run and we fail to restore that irrelevant sensitive parameter. The error is confusing, it will just be that we couldn't find a secret named RUNID-PARAM_NAME and it won't be immediately clear why that particular paramter wasn't persisted for that run, unless you realize that it didn't apply...

This only became a problem when I fixed how we remember parameters on an installation so that we remember ALL previously used parameters, and not just the paramters used on the last run. Before we never got into this situation where irrelevant parameters were on the installation, but now that can happen and we need to be more careful about how we copy paramters from an installation to a run.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>