Pass pre-computed Run when executing bundle
Update our CNAB runtime implementation to accept a pre-computed Run, which may have already been persisted in a pending state. We now execute it as defined and have moved more of the logic for how the bundle should be executed out of the runtime. This allows us to generate the runs ahead of time, store them on a workflow, and execute later.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>