Fix inconsistent bundle out-of-date detection
When we consider if a previously built bundle is out-of-date and should be rebuilt, we look at the porter.yaml file and also the set of mixins used by the bundle. We were storing those mixins and their installed version in a _map_, which doesn't sort consistently. So when a bundle used more than one mixin, there was always a chance that it would sort differently from the last build and trigger a rebuild.

I have updated how we compare the mixins used by the old/new bundle so that we always sort them in alphabetical order.

I have also added a bunch of debug logging and trace data so that it's easier to troubleshoot a bundle being incorrectly flagged as out-of-date next time. 😅

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: Aleksey Barabanov <alekseybb@gmail.com>