Refactor secret source data structures
As I was working on advanced dependencies, I kept running into confusion about what some of these structs and fields represented. For example, a source mapping had both a Value field (representing the resolved value) and a Source.Value representing a hint to the secret strategy on how to resolve the value of parameter or credential.

I have renamed the data structures and fields a bit to better align with how they are used.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: Aleksey Barabanov <alekseybb@gmail.com>