Fix binding configuration file settings to command flags
When we were binding viper values to the cobra flags, it was occuring before the config file was loaded, so if you configured a flag in the config file, it wasn't persisting the correct value to the variable bound to the flag.

When the flag was not bound to the global porter Config.Data, such as porter build --no-lint, the configuration file values were not applied. This is the regression describe in #2735 which was caused by apply cobra values to viper too soon (before the config file was loaded).

I have split our viper configuration into a three step process:

1. Configure the viper instance, for example binding custom open telemetry environment variables.
2. Bind viper to the cobra flags, AFTER the config file has been loaded.
3. Apply the consolidated configuration in viper (which now contains both cobra flags, env vars and the config file) back to the global config in Config.Data

This ensures that viper has all relevant configuration loaded before binding its value back to the original command flags (i.e. our variables in code).

Fixes #2735

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: Aleksey Barabanov <alekseybb@gmail.com>