Fix test to work on dev machines
I have updated this test to only check for the desired string, and not attempt to match the entire debug output generated by porter, which can contain stray debug lines depending on the dev env.
For example, on my machine I have telemetry enabled but when the backing otel collector is down, porter logs debug lines that it can't connect. The test is failing locally for me cause the test is requiring that ONLY that one line is printed, instead of just checking if the line was printed at all (more accurate and less fragile).

Since the test is intended to check that we are ****'ing out sensitive words, it doesn't need to be that strict.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: Aleksey Barabanov <alekseybb@gmail.com>