Do not pin invocation image to a specific debian release
Use stable instead of a specific debian release, so that when the release is archived, we don't end up with an unbuildable bundle. If a user wants to pin to a base image, they should use a custom Dockerfile. Porter should default to a release (stable) that will always work, so that porter doesn't need to be recompiled or a new version installed because a debian release was archived.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>