Run signing integration test
In order parallelize integration tests, the Github workflow doesn't use
Mage to run all integration tests directly. Instead each integration test
file is run in different jobs, requiring new integration test files to
be manually added to the workflow.
The integration test for signing images was missing and is added in this
PR.

Signed-off-by: Kim Christensen <kimworking@gmail.com>