Fix canary build of the server image
Building and publishing the server image during a canary build, did not
use the canary tag, instead it used the version of the build, e.g.,
v1.1.0-7-gd0387ecb. It also means that on each canary build a new tag
is used when publishing the image.

By checking if we are doing a release of a tagged version or a canary
build we can use the correct tag.

Signed-off-by: Kim Christensen <kimworking@gmail.com>