Add installation name to `porter upgrade` command in Quickstart (#3182)
When following the Quickstart for the first time, I received an error in the "Upgrade the Installation" section:

```
$ porter upgrade
no bundle specified. Either an installation name, --reference, --file or --cnab-file must be specified or the current directory must contain a porter.yaml file
```

I believe the intention was for this command to include an installation name. It does succeed when running `porter upgrade porter-hello` so I've updated the command

Signed-off-by: Eric Herrmann <26093534+egherrmann@users.noreply.github.com>
Co-authored-by: schristoff <28318173+schristoff@users.noreply.github.com>