Support custom templating in image section during lint
Images are validated for valid digests during linting. If the image
digest is templating and no default value exist in the bundle definition
the linting will fail, because the digest will be invalid.

One gotcha here is that if only part of the digest is templated linting
will still fail.

Signed-off-by: Kim Christensen <kimworking@gmail.com>