fix: special resolution of file params
Remove parameter resolution short-cut in `loadParameterSets` so that the
name (or value) of file parameters are resolved in the same way as other
parameters, unless the parameter strategy is `path`. Later, in
`finalizeParameters`, `getUnconvertedValueFromRaw` will perform special
handling to either load the file or pass the data along assuming it is
base64 encoded. 

Signed-off-by: Erik Cederberg <erik.cederberg@sectra.com>