Update MongoDB Docker Storage Plugin to use MongoDB 8 (#3357)
* Update MongoDB Docker image version from 4.0-xenial to 8.0-noble

Signed-off-by: Kim Christensen <kimworking@gmail.com>

* Add documentation for migrating MongoDB data to version 8.0

This new document outlines the steps required to upgrade existing MongoDB data stored in Docker from versions prior to Porter v1.3.0 to the new MongoDB v8.0. It includes instructions for backing up data, performing the upgrade through multiple version transitions, and a rollback procedure if needed.

Signed-off-by: Kim Christensen <kimworking@gmail.com>

* Update MongoDB Docker image references to version 8.0-noble in installation scripts and documentation

Signed-off-by: Kim Christensen <kimworking@gmail.com>

* Add delay in TestRoundTripDataOverGRPC to ensure proper indexing of documents

This change introduces a 100 millisecond sleep in the test to allow for all documents to be properly indexed before proceeding with the aggregation results check. This aims to improve the reliability of the test outcomes.

Signed-off-by: Kim Christensen <kimworking@gmail.com>

* chore: Set Bash modifiers

Signed-off-by: Kim Christensen <kimworking@gmail.com>

---------

Signed-off-by: Kim Christensen <kimworking@gmail.com>